Feature/merge readonly calendar (#70)
