--- Xamarin.Essentials/Calendar/CalendarRequest.ios.cs
@@ -1,12 +1,0 @@
-﻿using System;
-using EventKit;
-
-namespace Xamarin.Essentials
-{
-    static class CalendarRequest
-    {
-        static readonly Lazy<EKEventStore> EventStore = new Lazy<EKEventStore>(() => new EKEventStore());
-
-        public static EKEventStore Instance => EventStore.Value;
-    }
-}
